fix: address CI linting and type checking errors
- Replace print statements with logging in matlab_quality_check.py (Ruff T201)
- Fix progress_value type: change from int to float in update_status method
- Remove unused type: ignore comments in launch_integrated.py and Data_Processor_r0.py
- Fix return type annotations in performance_benchmark.py functions
- Add None check for time_col in performance_benchmark.py
- Add type casts in cli.py to resolve MyPy object type issues
- Fix import in launch_converter.py (use IntegratedCSVProcessorApp)
- Add pytest to mypy.ini ignore_missing_imports

+++ python/data_processor/Data_Processor_r0.py
@@ -6262,7 +6262,7 @@ This section helps you manage which signals (columns) to process from your files
         self,
         message: str,
         show_progress: bool = False,
-        progress_value: int = 0,
+        progress_value: float = 0.0,
         progress_text: str = "",
     ) -> None:
         """Update status bar with optional progress tracking."""
@@ -6498,7 +6498,7 @@ This section helps you manage which signals (columns) to process from your files
                         label=label,
                         color=color,
                         linewidth=line_width,
-                        **style_args,  # type: ignore[arg-type]
+                        **style_args,
                     )
 
                     # Show both raw and filtered if requested
